+++ src/Allocator.cpp
@@ -1,7 +1,7 @@
 #include "Allocator.h"
 #include <stdlib.h>     /* malloc, free */
 
-Allocator::Allocator(const uint32_t totalSize){
+Allocator::Allocator(const long totalSize){
 	m_totalSize = totalSize;
 	m_start_ptr = malloc(m_totalSize);
 }
